back-end/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
azure-common = "1.1.28"
azure-core = "1.30.1"
azure-identity = "1.16.0"
azure-search-documents = "11.4.0"
langchain = "0.1.17"
langchain-community = "0.0.36"
langchain-core = "0.1.49"
langchain-openai = "0.1.5"
langchain-text-splitters = "0.0.1"
msal = "1.28.0"
openai = "1.25.1"
pydantic = "2.7.1"
python-dotenv = "1.0.1"
tiktoken = "0.6.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
pypdf4 = "^1.27.0"
python-multipart = "^0.0.9"
youtube-transcript-api = "^0.6.2"
pytube = "^15.0.0"
azure-storage-blob = "^12.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
